Widen the lower bound of a measurement to 0.1ns because 2ns was insufficient for measuring static method calls on my MacBook Pro.
git-svn-id: https://caliper.googlecode.com/svn/trunk@189 62ec9553-1d4c-0410-be64-ad5e8ed6b23e

+++ src/com/google/caliper/UserException.java
@@ -124,8 +124,10 @@ public abstract class UserException extends RuntimeException {
   }
 
   public static class RuntimeOutOfRangeException extends ErrorInUsageException {
-    public RuntimeOutOfRangeException(long nanosPerExecution, long lowerBound, long upperBound) {
-      super("Runtime " + nanosPerExecution + "ns/rep out of range " + lowerBound + "-" + upperBound);
+    public RuntimeOutOfRangeException(
+        double nanosPerExecution, double lowerBound, double upperBound) {
+      super("Runtime " + nanosPerExecution + "ns/rep out of range "
+          + lowerBound + "-" + upperBound);
     }
   }
